by gacp

+++ Module-09/ex02/PmergeMe.cpp
@@ -6,7 +6,7 @@
 /*   By: youmoukh <youmoukh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/17 13:56:14 by youmoukh          #+#    #+#             */
-/*   Updated: 2024/10/19 16:14:06 by youmoukh         ###   ########.fr       */
+/*   Updated: 2024/10/19 16:15:23 by youmoukh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ void PmergeMe::OPerationParse(const std::string& Data_Entry)
 
 	for (size_t i = 0x0; i < Data_Entry.size(); i++)
 	{
-		int j = 0x0;
+		size_t j = 0x0;
 		while (j < Data_Entry.size())
 		{
 			if (isspace(Data_Entry[j]))
